+++ include/mujoco/mujoco.h
@@ -1411,11 +1411,8 @@ MJAPI void mju_taskJoin(mjTask* task);
 MJAPI mjsElement* mjs_attach(mjsElement* parent, const mjsElement* child,
                              const char* prefix, const char* suffix);
 
-// Delete body and descendants from mjSpec, remove all references, return 0 on success.
-MJAPI int mjs_detachBody(mjSpec* s, mjsBody* b);
-
-// Delete default class and descendants from mjSpec, remove all references, return 0 on success.
-MJAPI int mjs_detachDefault(mjSpec* s, mjsDefault* d);
+// Detach but not delete object corresponding to the given element, return 0 on success.
+MJAPI int mjs_detach(mjSpec* spec, mjsElement* element);
 
 //---------------------------------- Tree elements -------------------------------------------------
 
@@ -1444,7 +1441,7 @@ MJAPI mjsLight* mjs_addLight(mjsBody* body, const mjsDefault* def);
 MJAPI mjsFrame* mjs_addFrame(mjsBody* body, mjsFrame* parentframe);
 
 // Delete object corresponding to the given element, return 0 on success.
-MJAPI int mjs_delete(mjsElement* element);
+MJAPI int mjs_delete(mjSpec* spec, mjsElement* element);
 
 
 //---------------------------------- Non-tree elements ---------------------------------------------